add unit tests for category management and expense handling
