+++ src/nuxt-finanztracker_/test/unit/frontend/Test_ausgabe/Kontobewegung.expense.popup.cancel.spec.ts
@@ -1,6 +1,8 @@
+// test/unit/frontend/Test_ausgabe/Kontobewegung.expense.popup.cancel.spec.ts
+// Frontend Unit Test – Kontobewegung Ausgabe Popup abbrechen
 import { describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
-import Kontobewegung from '../../../pages/kontobewegung.vue'
+import Kontobewegung from '../../../../pages/kontobewegung.vue'
 
 describe('Kontobewegung – Ausgabe abbrechen', () => {
   it('setzt showExpenseModal auf false', async () => {
