+++ src/nuxt-finanztracker_/test/unit/frontend/Test_ausgabe/Kontobewegung.expense.popup.delete.spec.ts
@@ -1,27 +1,31 @@
+// test/unit/frontend/Test_ausgabe/Kontobewegung.expense.popup.delete.spec.ts
+// Frontend Unit Test – Kontobewegung Delete Popup (Ausgabe)
+
 import { mount } from '@vue/test-utils'
 import { describe, it, expect } from 'vitest'
-import Kontobewegung from '../../../pages/kontobewegung.vue'
+import Kontobewegung from '../../../../pages/kontobewegung.vue'
 
-describe('Kontobewegung – Delete Popup (Einnahme)', () => {
+describe('Kontobewegung – Delete Popup (Ausgabe)', () => {
   it('öffnet und schließt das Delete-Popup', async () => {
     const wrapper = mount(Kontobewegung)
     const vm = wrapper.vm as any
 
-    // add a sample transaction to enable delete action
+    // setup a sample expense transaction
     vm.transactions = [
-      { id: 1, type: 'Einnahme', amount: '100 €' }
+      { id: 2, type: 'Ausgabe', amount: '50 €' }
     ]
 
-    // simulate opening the delete popup
+    // simulate opening the popup
     vm.showActionModal = true
     vm.actionMode = 'delete'
     await wrapper.vm.$nextTick()
-    // verify that the delete popup is shown
+
     expect(vm.showActionModal).toBe(true)
 
+    // simulate closing the popup
     vm.showActionModal = false
     await wrapper.vm.$nextTick()
-    // verify that the delete popup is closed
+
     expect(vm.showActionModal).toBe(false)
   })
 })
