+++ src/nuxt-finanztracker_/test/unit/frontend/Test_einnahme/Kontobewegung.income.popup.cancel.spec.ts
@@ -1,6 +1,9 @@
+// import Kontobewegung from '../../../../pages/kontobewegung.vue'
+// Frontend Unit Test – Kontobewegung Einnahme Popup abbrechen
+
 import { describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
-import Kontobewegung from '../../../pages/kontobewegung.vue'
+import Kontobewegung from '../../../../pages/kontobewegung.vue'
 
 describe('Kontobewegung – Einnahme abbrechen', () => {
   it('setzt showIncomeModal auf false', async () => {
