+++ src/nuxt-finanztracker_/test/unit/frontend/Test_einnahme/Kontobewegung.incomes.category.select.spec.ts
@@ -1,6 +1,8 @@
+// test/unit/frontend/Test_einnahme/Kontobewegung.incomes.category.select.spec.ts
+// Frontend Unit Test – Kontobewegung Kategorie-Auswahl (Einnahme)
 import { describe, it, expect } from 'vitest'
 import { mount } from '@vue/test-utils'
-import Kontobewegung from '../../../pages/kontobewegung.vue'
+import Kontobewegung from '../../../../pages/kontobewegung.vue'
 
 describe('Kontobewegung – Kategorie-Auswahl (Einnahme)', () => {
   it('setzt showIncomeModal auf true', async () => {
